FakeInsightAsyncJob, FakeInsightAsyncJobManager - extract
