+++ airbyte-integrations/connectors/source-facebook-marketing/unit_tests/helpers.py
@@ -26,3 +26,29 @@ def read_incremental(stream_instance: Stream, stream_state: MutableMapping[str, 
     stream_state.clear()
     stream_state.update(stream_instance.state)
     return records
+
+
+class FakeInsightAsyncJobManager:
+    def __init__(self, jobs, **kwargs):
+        self.jobs = jobs
+
+    def completed_jobs(self):
+        yield from self.jobs
+
+
+class FakeInsightAsyncJob:
+    updated_insights = {}
+
+    @classmethod
+    def update_insight(cls, date_start, updated_time):
+        cls.updated_insights[date_start] = updated_time
+
+    def __init__(self, interval, **kwargs):
+        self.interval = interval
+
+    def get_result(self):
+        return [self]
+
+    def export_all_data(self):
+        date_start = str(self.interval.start)
+        return {"date_start": date_start, "updated_time": self.updated_insights.get(date_start, date_start)}
