+++ airbyte-integrations/connectors/source-facebook-marketing/unit_tests/test_base_insight_streams.py
@@ -8,7 +8,7 @@ import pendulum
 import pytest
 import source_facebook_marketing.streams.base_insight_streams
 from airbyte_cdk.models import SyncMode
-from helpers import read_full_refresh, read_incremental
+from helpers import FakeInsightAsyncJob, FakeInsightAsyncJobManager, read_full_refresh, read_incremental
 from pendulum import duration
 from source_facebook_marketing.streams import AdsInsights
 from source_facebook_marketing.streams.async_job import AsyncJob, InsightAsyncJob
@@ -298,20 +298,8 @@ class TestBaseInsightsStream:
         end_date = pendulum.parse("2020-04-10")
         monkeypatch.setattr(pendulum, "today", mocker.MagicMock(return_value=pendulum.parse("2020-04-01")))
 
-        class InsightAsyncJob:
-            def __init__(self, interval, **kwargs):
-                self.interval = interval
-
-        monkeypatch.setattr(source_facebook_marketing.streams.base_insight_streams, "InsightAsyncJob", InsightAsyncJob)
-
-        class InsightAsyncJobManager:
-            def __init__(self, jobs, **kwargs):
-                self.jobs = jobs
-
-            def completed_jobs(self):
-                yield from self.jobs
-
-        monkeypatch.setattr(source_facebook_marketing.streams.base_insight_streams, "InsightAsyncJobManager", InsightAsyncJobManager)
+        monkeypatch.setattr(source_facebook_marketing.streams.base_insight_streams, "InsightAsyncJob", FakeInsightAsyncJob)
+        monkeypatch.setattr(source_facebook_marketing.streams.base_insight_streams, "InsightAsyncJobManager", FakeInsightAsyncJobManager)
 
         state = {
             AdsInsights.cursor_field: "2020-03-20",
@@ -338,29 +326,8 @@ class TestBaseInsightsStream:
         end_date = pendulum.parse("2020-04-10")
         monkeypatch.setattr(pendulum, "today", mocker.MagicMock(return_value=pendulum.parse("2020-04-01")))
 
-        class InsightAsyncJob:
-            date_start_to_updated_time = {}
-
-            def __init__(self, interval, **kwargs):
-                self.interval = interval
-
-            def get_result(self):
-                return [self]
-
-            def export_all_data(self):
-                date_start = str(self.interval.start)
-                return {"updated_time": self.date_start_to_updated_time.get(date_start, date_start), "date_start": date_start}
-
-        monkeypatch.setattr(source_facebook_marketing.streams.base_insight_streams, "InsightAsyncJob", InsightAsyncJob)
-
-        class InsightAsyncJobManager:
-            def __init__(self, jobs, **kwargs):
-                self.jobs = jobs
-
-            def completed_jobs(self):
-                yield from self.jobs
-
-        monkeypatch.setattr(source_facebook_marketing.streams.base_insight_streams, "InsightAsyncJobManager", InsightAsyncJobManager)
+        monkeypatch.setattr(source_facebook_marketing.streams.base_insight_streams, "InsightAsyncJob", FakeInsightAsyncJob)
+        monkeypatch.setattr(source_facebook_marketing.streams.base_insight_streams, "InsightAsyncJobManager", FakeInsightAsyncJobManager)
 
         AdsInsights.INSIGHTS_LOOKBACK_PERIOD = pendulum.duration(days=20)
         stream = AdsInsights(api=api, start_date=start_date, end_date=end_date)
@@ -384,7 +351,8 @@ class TestBaseInsightsStream:
         assert state == {"date_start": "2020-04-02", "slices": [], "time_increment": 1}
 
         monkeypatch.setattr(pendulum, "today", mocker.MagicMock(return_value=pendulum.parse("2020-04-03")))
-        InsightAsyncJob.date_start_to_updated_time = {"2020-03-27": "2020-04-02", "2020-03-28": "2020-04-03"}
+        FakeInsightAsyncJob.update_insight("2020-03-27", "2020-04-02")
+        FakeInsightAsyncJob.update_insight("2020-03-28", "2020-04-03")
 
         records = read_incremental(stream, state)
         assert records == [
